add DefaultArgumentValues in config
